requirements.txt
click==8.1.7
isodate==0.6.1
lxml==5.2.2
numpy==2.0.0
pandas==2.2.2
pillow==10.3.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-docx==1.1.2
pytz==2024.1
rdflib==7.0.0
six==1.16.0
typing_extensions==4.12.2
tzdata==2024.1
